+++ TypeContractor/Output/OutputType.cs
@@ -1,9 +1,18 @@
-﻿using System.Globalization;
+using System.Globalization;
 using System.Text;
 
 namespace TypeContractor.Output;
 
-public record OutputType(string Name, string FullName, string FileName, ContractedType ContractedType, bool IsEnum, ICollection<OutputProperty>? Properties, ICollection<OutputEnumMember>? EnumMembers)
+public record OutputType(
+	string Name,
+	string FullName,
+	string FileName,
+	ContractedType ContractedType,
+	bool IsEnum,
+	bool IsGeneric,
+	ICollection<DestinationType> GenericTypeArguments,
+	ICollection<OutputProperty>? Properties,
+	ICollection<OutputEnumMember>? EnumMembers)
 {
 	public override string ToString()
 	{